OKAPI-947 revert changes to _tenantPermissions
