+++ okapi-core/src/main/java/org/folio/okapi/managers/TenantManager.java
@@ -24,8 +24,6 @@ import org.folio.okapi.bean.InterfaceDescriptor;
 import org.folio.okapi.bean.ModuleDescriptor;
 import org.folio.okapi.bean.ModuleInstance;
 import org.folio.okapi.bean.PermissionList;
-import org.folio.okapi.bean.PermissionListV1;
-import org.folio.okapi.bean.PermissionListV2;
 import org.folio.okapi.bean.RoutingEntry;
 import org.folio.okapi.bean.Tenant;
 import org.folio.okapi.bean.TenantDescriptor;
@@ -283,7 +281,7 @@ public class TenantManager implements Liveness {
     if (mdFrom == null && mdTo == null) {
       return Future.succeededFuture("");
     }
-    return invokePermissions(tenant, options, mdTo, mdFrom, pc)
+    return invokePermissions(tenant, options, mdTo, pc)
         .compose(x -> invokeTenantInterface(tenant, options, mdFrom, mdTo, pc))
         .compose(x -> invokePermissionsPermMod(tenant, options, mdFrom, mdTo, pc))
         .compose(x -> commitModuleChange(tenant, mdFrom, mdTo, pc))
@@ -399,12 +397,11 @@ public class TenantManager implements Liveness {
    * @param tenant tenant
    * @param options install options
    * @param mdTo module to
-   * @param mdFrom module from
    * @param pc proxy content
    * @return Future
    */
   private Future<Void> invokePermissions(Tenant tenant, TenantInstallOptions options,
-      ModuleDescriptor mdTo, ModuleDescriptor mdFrom, ProxyContext pc) {
+      ModuleDescriptor mdTo, ProxyContext pc) {
     if (mdTo == null || !options.getInvoke()
         || mdTo.getSystemInterface("_tenantPermissions") != null) {
       return Future.succeededFuture();
@@ -413,7 +410,7 @@ public class TenantManager implements Liveness {
       if (md == null) {
         return Future.succeededFuture();
       }
-      return invokePermissionsForModule(tenant, mdTo, mdFrom, md, pc);
+      return invokePermissionsForModule(tenant, mdTo, md, pc);
     });
   }
 
@@ -437,7 +434,7 @@ public class TenantManager implements Liveness {
     // enabling permissions module.
     return findSystemInterface(tenant, "_tenantPermissions")
         .compose(res -> loadPermissionsForEnabled(tenant, mdTo, pc))
-        .compose(res -> invokePermissionsForModule(tenant, mdTo, mdFrom, mdTo, pc));
+        .compose(res -> invokePermissionsForModule(tenant, mdTo, mdTo, pc));
   }
 
   /**
@@ -454,7 +451,7 @@ public class TenantManager implements Liveness {
     Future<Void> future = Future.succeededFuture();
     for (String mdid : tenant.listModules()) {
       future = future.compose(x -> moduleManager.get(mdid)
-          .compose(md -> invokePermissionsForModule(tenant, md, md, permsModule, pc)));
+          .compose(md -> invokePermissionsForModule(tenant, md, permsModule, pc)));
     }
     return future;
   }
@@ -530,7 +527,7 @@ public class TenantManager implements Liveness {
       TenantInstallOptions options = new TenantInstallOptions();
       String fromVer = new ModuleId(enver).getSemVer().toString();
       ModuleDescriptor mdFrom = InternalModule.moduleDescriptor(fromVer);
-      return invokePermissions(tenant, options, md, mdFrom, null).compose(x ->
+      return invokePermissions(tenant, options, md, null).compose(x ->
           updateModuleCommit(tenant, moduleFrom, moduleTo));
     });
   }
@@ -645,7 +642,7 @@ public class TenantManager implements Liveness {
   }
 
   private Future<Void> invokePermissionsForModule(Tenant tenant, ModuleDescriptor mdTo,
-      ModuleDescriptor mdFrom, ModuleDescriptor permsModule, ProxyContext pc) {
+      ModuleDescriptor permsModule, ProxyContext pc) {
 
     logger.debug("Loading permissions for {} (using {})", mdTo.getName(), permsModule.getName());
     String moduleTo = mdTo.getId();
@@ -653,19 +650,9 @@ public class TenantManager implements Liveness {
     InterfaceDescriptor permInt = permsModule.getSystemInterface("_tenantPermissions");
     String permIntVer = permInt.getVersion();
     if (permIntVer.equals("1.0")) {
-      pl = new PermissionListV1(moduleTo, mdTo.getPermissionSets());
-    } else if (permIntVer.equals("1.1")) {
-      pl = new PermissionListV1(moduleTo, mdTo.getExpandedPermissionSets());
-    } else if (permIntVer.equals("2.0")) {
-      if (mdFrom != null) {
-        pl = new PermissionListV2(moduleTo, mdFrom.getId(), mdTo.getExpandedPermissionSets(),
-            mdFrom.getExpandedPermissionSets());
-      } else {
-        pl = new PermissionListV2(moduleTo, mdTo.getExpandedPermissionSets());
-      }
+      pl = new PermissionList(moduleTo, mdTo.getPermissionSets());
     } else {
-      return Future.failedFuture(new OkapiError(ErrorType.USER,
-          "Unknown version of _tenantPermissions interface in use " + permIntVer + "."));
+      pl = new PermissionList(moduleTo, mdTo.getExpandedPermissionSets());
     }
     String pljson = Json.encodePrettily(pl);
     logger.debug("tenantPerms Req: {}", pljson);
