+++ okapi-core/src/test/java/org/folio/okapi/ProxyTest.java
@@ -253,12 +253,7 @@ public class ProxyTest {
           } else if (p.startsWith("/permissionscall")) {
             JsonObject permObject = new JsonObject(buf);
             if (timerTenantPermissionsStatus == 200) {
-              String moduleToId = permObject.getString("moduleToId");
-              if (moduleToId != null) {
-                timerPermissions.put(moduleToId, permObject.getJsonArray("permsTo"));
-              } else {
-                timerPermissions.put(permObject.getString("moduleId"), permObject.getJsonArray("perms"));
-              }
+              timerPermissions.put(permObject.getString("moduleId"), permObject.getJsonArray("perms"));
             }
             response.setStatusCode(timerTenantPermissionsStatus);
             response.end("timer permissions response");
@@ -3847,8 +3842,8 @@ public class ProxyTest {
 
     setupBasicTenant(tenant);
 
-    // test _tenantpermissions 1.0 vs 1.1 vs 2.0
-    for (String tenantPermissionsVersion : Arrays.asList("1.0", "1.1", "2.0")) {
+    // test _tenantpermissions 1.0 vs 1.1
+    for (String tenantPermissionsVersion : Arrays.asList("1.0", "1.1")) {
       timerPermissions.clear();
       setupBasicModule(tenant, moduleId, tenantPermissionsVersion, true, true);
       setupBasicAuth(tenant, authModuleId);
